refactor(route): 아티스트카드-이미지 클릭 시 경로 수정 및 추천 카드 폴더 이동
